Move the debug_ui shards to egui
